recko_backend/AIS/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-registration==0.5.6
django-rest-swagger==2.2.0
django-social-auth==0.7.28
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
ecdsa==0.14.1
enum-compat==0.0.3
enum34==1.1.10
eventlet==0.21.0
future==0.18.2
greenlet==1.0.0
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
iniconfig==1.1.1
intuit-oauth==1.1.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
mock==4.0.3
msgpack==0.6.2
oauth2==1.9.0.post1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytest==6.2.2
pytest-cov==2.5.0
python-coveralls==2.9.3
python-dateutil==2.8.1
python-jose==3.2.0
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
pyxero==0.9.2
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
simplejson==3.17.2
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
xero-python==1.4.0
yapf==0.30.0
